# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLMNightRun",
    "description": "멀티 LLM 통합 자동화 플랫폼",
    "version": "0.1.0",
    "x-logo": {
      "url": "/static/logo.png",
      "altText": "LLMNightRun"
    }
  },
  "paths": {
...
    "/api/v2/llm/generate": {
      "post": {
        "tags": [
          "LLM"
        ],
        "summary": "LLM 텍스트 생성",
        "description": "LLM에 프롬프트를 전송하고 텍스트를 생성합니다.\n\nArgs:\n    request: LLM 프롬프트 요청 데이터\n    controller: LLM 컨트롤러 인스턴스\n    \nReturns:\n    생성된 텍스트와 메타데이터",
        "operationId": "generate_text_api_v2_llm_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMPromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMPromptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "not_found",
                  "message": "요청한 리소스를 찾을 수 없습니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/llm/models": {
      "get": {
        "tags": [
          "LLM"
        ],
        "summary": "사용 가능한 LLM 모델 목록 조회",
        "description": "사용 가능한 LLM 모델 목록을 반환합니다.\n\nArgs:\n    llm_type: LLM 유형 (선택 사항)\n    controller: LLM 컨트롤러 인스턴스\n    \nReturns:\n    사용 가능한 모델 목록",
        "operationId": "get_models_api_v2_llm_models_get",
        "parameters": [
          {
            "name": "llm_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LLMType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Llm Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Models Api V2 Llm Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "not_found",
                  "message": "요청한 리소스를 찾을 수 없습니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMPromptRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "LLM에 전송할 프롬프트"
          },
          "llm_type": {
            "$ref": "#/components/schemas/LLMType",
            "description": "사용할 LLM 유형"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "사용할 모델 이름 (선택 사항)"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "description": "온도 (0~1) (선택 사항)"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens",
            "description": "최대 토큰 수 (선택 사항)"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt",
            "description": "시스템 프롬프트 (선택 사항)"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "llm_type"
        ],
        "title": "LLMPromptRequest",
        "description": "LLM 프롬프트 요청 모델"
      },
      "LLMPromptResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "요청 성공 여부"
          },
          "data": {
            "type": "object",
            "title": "Data",
            "description": "응답 데이터"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "응답 메시지"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "title": "LLMPromptResponse",
        "description": "LLM 프롬프트 응답 모델"
      },
      "LLMType": {
        "type": "string",
        "enum": [
          "openai_api",
          "openai_web",
          "claude_api",
          "claude_web",
          "local_llm",
          "custom_api",
          "manual"
        ],
        "title": "LLMType",
        "description": "LLM 유형 정의"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
  "servers": [
    {
      "url": "",
      "description": "개발 서버"
    }
  ]
}